Conflicts:
	lucene/CHANGES.txt
	lucene/codecs/src/java/org/apache/lucene/codecs/memory/DirectDocValuesProducer.java
	lucene/codecs/src/java/org/apache/lucene/codecs/memory/MemoryDocValuesProducer.java
	lucene/codecs/src/java/org/apache/lucene/codecs/simpletext/SimpleTextDocValuesReader.java
	lucene/core/src/java/org/apache/lucene/analysis/Token.java
	lucene/core/src/java/org/apache/lucene/codecs/lucene42/Lucene42DocValuesProducer.java
	lucene/core/src/java/org/apache/lucene/codecs/lucene45/Lucene45DocValuesProducer.java
	lucene/core/src/java/org/apache/lucene/index/BinaryDocValues.java
	lucene/core/src/java/org/apache/lucene/index/DocInverterPerField.java
	lucene/core/src/java/org/apache/lucene/index/DocTermOrds.java
	lucene/core/src/java/org/apache/lucene/index/DocValues.java
	lucene/core/src/java/org/apache/lucene/index/DocumentsWriterPerThread.java
	lucene/core/src/java/org/apache/lucene/index/DocumentsWriterPerThreadPool.java
	lucene/core/src/java/org/apache/lucene/index/FlushPolicy.java
	lucene/core/src/java/org/apache/lucene/index/LiveIndexWriterConfig.java
	lucene/core/src/java/org/apache/lucene/index/MultiDocValues.java
	lucene/core/src/java/org/apache/lucene/index/SegmentMerger.java
	lucene/core/src/java/org/apache/lucene/index/SegmentReader.java
	lucene/core/src/java/org/apache/lucene/index/SingletonSortedSetDocValues.java
	lucene/core/src/java/org/apache/lucene/index/SlowCompositeReaderWrapper.java
	lucene/core/src/java/org/apache/lucene/index/TermsHashPerField.java
	lucene/core/src/java/org/apache/lucene/search/FieldCacheImpl.java
	lucene/core/src/java/org/apache/lucene/store/CompoundFileDirectory.java
	lucene/core/src/java/org/apache/lucene/store/FSDirectory.java
	lucene/core/src/java/org/apache/lucene/store/LockStressTest.java
	lucene/core/src/java/org/apache/lucene/store/NativeFSLockFactory.java
	lucene/core/src/java/org/apache/lucene/store/RateLimitedIndexOutput.java
	lucene/core/src/java/org/apache/lucene/util/RamUsageEstimator.java
	lucene/core/src/test/org/apache/lucene/analysis/TestToken.java
	lucene/core/src/test/org/apache/lucene/index/Test2BSortedDocValues.java
	lucene/core/src/test/org/apache/lucene/index/TestBackwardsCompatibility.java
	lucene/core/src/test/org/apache/lucene/index/TestIndexWriter.java
	lucene/core/src/test/org/apache/lucene/index/TestIndexWriterExceptions.java
	lucene/core/src/test/org/apache/lucene/index/TestIndexWriterExceptions2.java
	lucene/core/src/test/org/apache/lucene/index/TestIndexWriterOutOfMemory.java
	lucene/core/src/test/org/apache/lucene/index/TestIndexWriterThreadsToSegments.java
	lucene/core/src/test/org/apache/lucene/index/TestTieredMergePolicy.java
	lucene/core/src/test/org/apache/lucene/store/TestRateLimiter.java
	lucene/facet/src/java/org/apache/lucene/facet/taxonomy/DocValuesOrdinalsReader.java
	lucene/ivy-versions.properties
	lucene/replicator/src/test/org/apache/lucene/replicator/http/HttpReplicatorTest.java
	lucene/test-framework/src/java/org/apache/lucene/codecs/cranky/CrankyDocValuesFormat.java
	lucene/test-framework/src/java/org/apache/lucene/codecs/cranky/CrankyStoredFieldsFormat.java
	lucene/test-framework/src/java/org/apache/lucene/index/BaseDocValuesFormatTestCase.java
	lucene/test-framework/src/java/org/apache/lucene/index/BaseIndexFileFormatTestCase.java
	lucene/test-framework/src/java/org/apache/lucene/util/LuceneTestCase.java
	solr/CHANGES.txt
	solr/contrib/extraction/ivy.xml
	solr/core/src/java/org/apache/solr/cloud/ElectionContext.java
	solr/core/src/java/org/apache/solr/handler/FieldAnalysisRequestHandler.java
	solr/core/src/java/org/apache/solr/handler/MoreLikeThisHandler.java
	solr/core/src/java/org/apache/solr/handler/component/DebugComponent.java
	solr/core/src/java/org/apache/solr/handler/component/QueryComponent.java
	solr/core/src/java/org/apache/solr/request/DocValuesFacets.java
	solr/core/src/java/org/apache/solr/schema/IndexSchema.java
	solr/core/src/java/org/apache/solr/schema/ManagedIndexSchema.java
	solr/core/src/java/org/apache/solr/search/CollapsingQParserPlugin.java
	solr/core/src/java/org/apache/solr/store/hdfs/HdfsDirectory.java
	solr/core/src/test/org/apache/solr/TestDistributedSearch.java
	solr/core/src/test/org/apache/solr/handler/FieldAnalysisRequestHandlerTest.java
	solr/core/src/test/org/apache/solr/handler/component/DistributedDebugComponentTest.java
	solr/solrj/src/java/org/apache/solr/client/solrj/impl/LBHttpSolrServer.java
	solr/solrj/src/test/org/apache/solr/client/solrj/impl/CloudSolrServerTest.java

+++ lucene/test-framework/src/java/org/apache/lucene/util/TestUtil.java
@@ -46,7 +46,7 @@ import java.util.zip.ZipFile;
 import org.apache.lucene.codecs.Codec;
 import org.apache.lucene.codecs.DocValuesFormat;
 import org.apache.lucene.codecs.PostingsFormat;
-import org.apache.lucene.codecs.lucene46.Lucene46Codec;
+import org.apache.lucene.codecs.lucene49.Lucene49Codec;
 import org.apache.lucene.codecs.perfield.PerFieldDocValuesFormat;
 import org.apache.lucene.codecs.perfield.PerFieldPostingsFormat;
 import org.apache.lucene.document.BinaryDocValuesField;
@@ -120,14 +120,16 @@ public final class TestUtil {
   }
 
   private static LinkedHashSet<File> rm(LinkedHashSet<File> unremoved, File... locations) {
-    for (File location : locations) {
-      if (location.exists()) {
-        if (location.isDirectory()) {
-          rm(unremoved, location.listFiles());
-        }
-
-        if (!location.delete()) {
-          unremoved.add(location);
+    if (locations != null) {
+      for (File location : locations) {
+        if (location != null && location.exists()) {
+          if (location.isDirectory()) {
+            rm(unremoved, location.listFiles());
+          }
+  
+          if (!location.delete()) {
+            unremoved.add(location);
+          }
         }
       }
     }
@@ -680,7 +682,7 @@ public final class TestUtil {
     if (LuceneTestCase.VERBOSE) {
       System.out.println("forcing postings format to:" + format);
     }
-    return new Lucene46Codec() {
+    return new Lucene49Codec() {
       @Override
       public PostingsFormat getPostingsFormatForField(String field) {
         return format;
@@ -698,7 +700,7 @@ public final class TestUtil {
     if (LuceneTestCase.VERBOSE) {
       System.out.println("forcing docvalues format to:" + format);
     }
-    return new Lucene46Codec() {
+    return new Lucene49Codec() {
       @Override
       public DocValuesFormat getDocValuesFormatForField(String field) {
         return format;
